+++ wemake_python_styleguide/compat/constants.py
@@ -2,9 +2,6 @@ import sys
 
 from typing_extensions import Final
 
-#: This indicates that we are running on python3.10+
-PY310: Final = sys.version_info >= (3, 10)
-
 #: This indicates that we are running on python3.11+
 PY311: Final = sys.version_info >= (3, 11)